fix: Allow more than 10 items in sourceData of menu button inside table widget (#25724)
